Agregar stub compartido de telegram y actualizar pruebas
