+++ tests/test_email.py
@@ -1,29 +1,13 @@
 import sys
 import os
 import importlib
+import tests.telegram_stub  # Registra las clases fake de telegram
 from types import ModuleType
 from pathlib import Path
 from sqlalchemy.orm import sessionmaker
 
 ROOT_DIR = Path(__file__).resolve().parents[1]
 sys.path.append(str(ROOT_DIR / "Sandy bot"))
-
-# Stub de telegram requerido por utils
-telegram_stub = ModuleType("telegram")
-class Message:
-    def __init__(self, text=""):
-        self.text = text
-class CallbackQuery:
-    def __init__(self, message=None):
-        self.message = message
-class Update:
-    def __init__(self, message=None, edited_message=None, callback_query=None):
-        self.message = message
-        self.edited_message = edited_message
-        self.callback_query = callback_query
-telegram_stub.Update = Update
-telegram_stub.Message = Message
-sys.modules.setdefault("telegram", telegram_stub)
 
 # Stub de dotenv
 dotenv_stub = ModuleType("dotenv")
